Unify thread utilities
Put the various thread utilities into one place. Write some relevant
documentation for an otherwise-empty package. Alter support annotation
to make that practical.
