+++ SpiNNaker-nmpiexec/pom.xml
@@ -49,6 +49,10 @@
 			<artifactId>SpiNNaker-nmpimodel</artifactId>
 		</dependency>
 		<dependency>
+			<groupId>${project.groupId}</groupId>
+			<artifactId>SpiNNaker-utils</artifactId>
+		</dependency>
+		<dependency>
 			<groupId>org.ini4j</groupId>
 			<artifactId>ini4j</artifactId>
 		</dependency>
